+++ app/templates/scripts/general/jquery.js
@@ -1,4 +1,0 @@
-import $ from 'jquery'
-
-// Make jQuery global for other plugins to consume
-window.$ = window.jQuery = $