+++ test/bootstrap.js
@@ -2,7 +2,7 @@ import helpers from 'yeoman-test'
 import assert from 'yeoman-assert'
 import path from 'path'
 
-describe('Bootstrap features', function () {
+describe('Bootstrap 4 features', function () {
   beforeEach(function () {
     return helpers.run(path.join(__dirname, '../app'))
       .withOptions({
@@ -12,13 +12,51 @@ describe('Bootstrap features', function () {
         projectName: 'Pixel2HTML',
         qtyScreens: 3,
         markupLanguage: 'html',
-        frontEndFramework: 'bootstrap'
+        frontEndFramework: 'bootstrap-4'
       })
       .toPromise()
   })
 
   it('Should list dependencies in package.json', function () {
-    assert.fileContent('package.json', /"gulp-sass"/)
+    assert.fileContent('package.json', /"bootstrap"/)
+  })
+
+  it('should exists a gulp routine', function () {
+    assert.file([
+      'gulp/tasks/styles.js',
+      'gulp/tasks/scripts.js'
+    ])
+  })
+
+  it('should exists vendor files', function () {
+    assert.file([
+      'src/assets/styles/vendor/vendor.scss',
+      'src/assets/js/index.js'
+    ])
+  })
+
+  it('should include bootstrap include', function () {
+    assert.fileContent('src/assets/styles/vendor/vendor.scss', /import "bootstrap\/scss\/bootstrap.scss";/)
+    assert.fileContent('src/assets/js/framework.js', /import 'bootstrap'/)
+  })
+})
+
+describe('Bootstrap 3 features', function () {
+  beforeEach(function () {
+    return helpers.run(path.join(__dirname, '../app'))
+      .withOptions({
+        'skip-install': true
+      })
+      .withPrompts({
+        projectName: 'Pixel2HTML',
+        qtyScreens: 3,
+        markupLanguage: 'html',
+        frontEndFramework: 'bootstrap-3'
+      })
+      .toPromise()
+  })
+
+  it('Should list dependencies in package.json', function () {
     assert.fileContent('package.json', /"bootstrap-sass"/)
   })
 
@@ -31,14 +69,14 @@ describe('Bootstrap features', function () {
 
   it('should exists vendor files', function () {
     assert.file([
-      'src/assets/styles/vendor.scss',
-      'src/assets/js/general/bootstrap.js'
+      'src/assets/styles/vendor/vendor.scss',
+      'src/assets/js/index.js'
     ])
   })
 
   it('should include bootstrap include', function () {
-    assert.fileContent('src/assets/styles/vendor.scss', /import "bootstrap-sass";/)
-    assert.fileContent('src/assets/js/general/index.js', /import '.\/bootstrap'/)
+    assert.fileContent('src/assets/styles/vendor/vendor.scss', /import "bootstrap-sass";/)
+    assert.fileContent('src/assets/js/framework.js', /require\('bootstrap-sass'\)/)
   })
 
   it('should include correct paths on config file', function () {