+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 2.2.0
+- HMR
+- CSS HMR
+- PurifyCSS
+- CriticalCSS
+
+### 2.1.0
+- Add Bootstrap 4 Beta as an Frontend Framework option
+- Decouple jQuery from bundle on production mode.
+- Keep alive the opened browser tab when you hit `npm run code`. Thanks Facebook!
+
+
+
 ### 2.0.0
 - Reduced number of total dependencies
 - Webpack integration to handle JS files